% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_open_chromatin_regions}
\alias{consensus_open_chromatin_regions}
\title{Consensus open chromatin regions dataset}
\format{
A data frame with 2892942 rows and 4 variables:
\describe{
  \item{\code{chromosome}}{character Chromosome location of peaks}
  \item{\code{start}}{integer 1-based start location of peaks}
  \item{\code{end}}{integer 1-based end location of peaks}
  \item{\code{strand}}{integer Placeholder for strand information} 
}
}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTrackUi?db=hg19&g=wgEncodeAwgDnaseMasterSites}
}
\usage{
consensus_open_chromatin_regions
}
\description{
A dataset containing a master track of DNaseI hypersensitive sites from 125 seperate cell types from ENCODE lifted from the hg19 to the hg38 human genome.
}
\keyword{datasets}
