% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode_v22_annotations}
\alias{gencode_v22_annotations}
\title{GENCODE v22 .gtf file}
\format{A data frame with 2563671 rows and 27 variables:
\describe{
  \item{\code{seqnames}}{character Chromosome location of genes}
  \item{\code{start}}{integer 1-based start location of genes}
  \item{\code{end}}{integer 1-based end location of genes}
  \item{\code{width}}{integer Total width of the region}
  \item{\code{strand}}{character Strand the region is contained on}
  \item{\code{source}}{character Havana or Ensembl annotation}
  \item{\code{type}}{character type of construct defined by the region}
  \item{\code{score}}{character NA}
  \item{\code{phase}}{integer NA}
  \item{\code{gene_id}}{character ENSG# ID followed by a single digit after a period}
  \item{\code{gene_type}}{character COLUMN_DESCRIPTION}
  \item{\code{gene_status}}{character COLUMN_DESCRIPTION}
  \item{\code{gene_name}}{character COLUMN_DESCRIPTION}
  \item{\code{level}}{character COLUMN_DESCRIPTION}
  \item{\code{havana_gene}}{character COLUMN_DESCRIPTION}
  \item{\code{transcript_id}}{character COLUMN_DESCRIPTION}
  \item{\code{transcript_type}}{character COLUMN_DESCRIPTION}
  \item{\code{transcript_status}}{character COLUMN_DESCRIPTION}
  \item{\code{transcript_name}}{character COLUMN_DESCRIPTION}
  \item{\code{tag}}{character COLUMN_DESCRIPTION}
  \item{\code{transcript_support_level}}{character COLUMN_DESCRIPTION}
  \item{\code{havana_transcript}}{character COLUMN_DESCRIPTION}
  \item{\code{exon_number}}{character COLUMN_DESCRIPTION}
  \item{\code{exon_id}}{character COLUMN_DESCRIPTION}
  \item{\code{ont}}{character COLUMN_DESCRIPTION}
  \item{\code{protein_id}}{character COLUMN_DESCRIPTION}
  \item{\code{ccdsid}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
gencode_v22_annotations
}
\description{
A dataset containing information for the GENCODE v22 (GRCh38.p2) .gtf file
}
\keyword{datasets}
