% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_chrom_sizes}
\alias{hg38_chrom_sizes}
\title{hg38_chrom_sizes}
\format{
A data frame with 455 rows and 2 variables:
\describe{
  \item{\code{chromosome}}{character Chromosome and scaffold names.}
  \item{\code{size}}{integer Size of each chromosome and scaffold in base pairs} 
}
}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg38/bigZips/}
}
\usage{
hg38_chrom_sizes
}
\description{
A dataset of chromosome and scaffold sizes for the hg38 build of the human genome
}
\keyword{datasets}
