% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm450_hg38_annotations}
\alias{hm450_hg38_annotations}
\title{DATASET_TITLE}
\format{A data frame with 485577 rows and 57 variables:
\describe{
  \item{\code{CpG_chrm}}{character Chromosome location of probes}
  \item{\code{CpG_beg}}{integer 1-based start location of peaks}
  \item{\code{CpG_end}}{integer 1-based end location of peaks}
  \item{\code{probe_strand}}{character Strand the probe is contained on}
  \item{\code{probeID}}{character Name of the probe in form: 'cg########'}
  \item{\code{address_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{address_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{channel}}{character COLUMN_DESCRIPTION}
  \item{\code{designType}}{character COLUMN_DESCRIPTION}
  \item{\code{nextBase}}{character COLUMN_DESCRIPTION}
  \item{\code{nextBaseRef}}{character COLUMN_DESCRIPTION}
  \item{\code{probeType}}{character COLUMN_DESCRIPTION}
  \item{\code{orientation}}{character COLUMN_DESCRIPTION}
  \item{\code{probeCpGcnt}}{integer COLUMN_DESCRIPTION}
  \item{\code{context35}}{integer COLUMN_DESCRIPTION}
  \item{\code{probeBeg}}{integer COLUMN_DESCRIPTION}
  \item{\code{probeEnd}}{integer COLUMN_DESCRIPTION}
  \item{\code{ProbeSeq_A}}{character COLUMN_DESCRIPTION}
  \item{\code{ProbeSeq_B}}{character COLUMN_DESCRIPTION}
  \item{\code{gene}}{character COLUMN_DESCRIPTION}
  \item{\code{gene_HGNC}}{character COLUMN_DESCRIPTION}
  \item{\code{chrm_A}}{character COLUMN_DESCRIPTION}
  \item{\code{beg_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{flag_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{mapQ_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{cigar_A}}{character COLUMN_DESCRIPTION}
  \item{\code{NM_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{chrm_B}}{character COLUMN_DESCRIPTION}
  \item{\code{beg_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{flag_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{mapQ_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{cigar_B}}{character COLUMN_DESCRIPTION}
  \item{\code{NM_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_chrm_A}}{character COLUMN_DESCRIPTION}
  \item{\code{wDecoy_beg_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_flag_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_mapQ_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_cigar_A}}{character COLUMN_DESCRIPTION}
  \item{\code{wDecoy_NM_A}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_chrm_B}}{character COLUMN_DESCRIPTION}
  \item{\code{wDecoy_beg_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_flag_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_mapQ_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{wDecoy_cigar_B}}{character COLUMN_DESCRIPTION}
  \item{\code{wDecoy_NM_B}}{integer COLUMN_DESCRIPTION}
  \item{\code{posMatch}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_mapping}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_typeINextBaseSwitch}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_rmsk15}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_sub40_copy}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_sub35_copy}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_sub30_copy}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_sub25_copy}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_snp5_common}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_snp5_GMAF1p}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_extBase}}{logical COLUMN_DESCRIPTION}
  \item{\code{MASK_general}}{logical COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
hm450_hg38_annotations
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
