% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_enhancer_regions}
\alias{consensus_enhancer_regions}
\title{Multiple cell type consensus enhancer region dataset}
\format{A data frame with 393892 rows and 4 variables:
\describe{
  \item{\code{chromosome}}{character Chromosome location of peaks}
  \item{\code{start}}{integer 1-based start location of peaks}
  \item{\code{end}}{integer 1-based end location of peaks}
  \item{\code{strand}}{integer Placeholder for strand information} 
}}
\source{
\url{https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/core_K27ac/jointModel/final/bed_hg38_lifted_over/ https://egg2.wustl.edu/roadmap/web_portal/chr_state_learning.html#exp_18state}
}
\usage{
consensus_enhancer_regions
}
\description{
A dataset containing consensus enhancer regions for numerous cell types. Based on chromHMM annotations and eRNA datasets.
}
\keyword{datasets}
